pyproject.toml
[project]
name = "backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["PyYAML", "sh"]

[project.scripts]
backup = "backup.main:main"
